requirements.txt
aioredis==1.3.1
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.10
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.11
django-celery-beat==2.2.1
django-extensions==3.1.5
django-timezone-field==4.2.3
django-widget-tweaks==1.4.12
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
mpmath==1.2.1
msgpack==1.0.3
Naked==0.1.31
numpy==1.19.5
packaging==21.3
Pillow==9.0.0
pip-autoremove==0.10.0
pkg_resources==0.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
redis==4.2.2
requests==2.27.1
RPi.GPIO==0.7.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
sympy==1.9
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zope.interface==5.4.0
